Don't overwrite builds made in the same day
If two builds were done in the same day and uploaded to the OTA server, the
second would overwrite the first, resulting in the first build download link
pointing to the second build. Fix this by putting builds in their own directory
with a timestamp. Provided two builds don't get made in the same second, this
should be relatively futureproof.

Also, calculate md5sum before upload, and after signing. If signed builds were
made, the contents of the zip were changed, but the md5sum was not recalculated
so signed OTAs would fail validation. Fix this by not relying at all on the
'prebuilt' md5, as md5summing is fairly inexpensive in the grand scheme of
things